interface for retrofit holder
