+++ acnsdk/src/main/java/com/arrow/acn/api/AcnApiServiceFactory.java
@@ -11,7 +11,7 @@
 package com.arrow.acn.api;
 
 
-import com.arrow.acn.api.common.RetrofitHolder;
+import com.arrow.acn.api.common.RetrofitHolderImpl;
 
 /**
  * Created by osminin on 9/21/2016.
@@ -22,7 +22,7 @@ public final class AcnApiServiceFactory {
     private static AcnApiService service;
 
     public static AcnApiService createAcnApiService() {
-        service = new AcnApiImpl(new RetrofitHolder());
+        service = new AcnApiImpl(new RetrofitHolderImpl());
         return service;
     }
